fastgpt/src/requirements.txt
bleach
fastapi
openai
python-fasthtml
sse-starlette
loguru